refactor: reduce INFO verbosity
We're currently emitting ~5GB of logs every 30 minutes, and a quick scan
through the logs shows the lines this PR changes to be the most frequent
(multiple times per second).

I don't believe any of these are important enough to be INFO, but if
needed, an appropriate log filter will bring them back.