fix: remove controller 64bit misalignment
I understand 32-bit architectures are no longer officially supported,
but this is a minor change.

Calling `atomic.AddUint64()` on `lastID` resulted in an alignment panic
in armv7, due to it not being 64bit aligned.

Declaring it first in the struct solved it.

See https://pkg.go.dev/sync/atomic#pkg-note-BUG fore more details