ci: enable tokio_console feature
Configures the release CI pipeline to generate a binary with
tokio-console support.

This drops the compile-time log filter to TRACE, which may have some
minor overhead. All our log filters in prod are debug or higher, so the
runtime filter will remove the TRACE events and no additional logs will
be emitted because of this change.