fix: handle LP lines with duplicated fields/tags
This commit allows the LP consumer to correctly handle line protocol
writes that duplicate one or more fields or tags within a single line:

           table v=2,bananas=42,v=3,platanos=24
                  ▲              ▲
                  └───────┬──────┘
                          │
                 duplicate field "v"

This change implements the following logic when processing fields:

    IF field is duplicated
        IF all duplicate field values are of the same data type
            use last occurrence ("last write wins")
        ELSE
            return an error

Any duplication of tags is rejected, and use of a field name as both a
field and a tag is remains forbidden (unchanged in this PR, a previously
agreed breaking change from TSM).

See https://github.com/influxdata/influxdb_iox/issues/4326 for context.