test(label): Invalid closure capture
Whilst working on cloud labels, I noticed that the label storage test
isn't correctly capturing the testScenario. This is because t.Parallel()
is in use, so only the "delete" scenario will only ever be tested.

The following changes fix that, by capturing the testScenario in a local
variable so that the test function can use that correctly.

As a drive-by, I fixed the error referencing the wrong expectation.