fix(tsm1): shrinking the time window for merging overlapping blocks
when scanning data with overlapping blocks, the current algorithm would
maximize time window(though we do have comments mentioning it *should*
shrink the window). If we have a block with a long time window
which overlaps with many other blocks, the merging operation would be
time consuming especially with string data type.
This change simply makes the code in line with original comments by
shrinking the window to merge.